COM-2149: mutualize stuffs
